Create new tmux sessions from project dirnames
The script `$RC/bin/tmux-session-from-dir` starts by listing all the
project directories inside my personal and work code directories and
feeds that into a `fzf` prompt that I use to select which one I'm
interested in, called $selected. After that, it creates a tmux-session
with $selected as the working directory and attaches to it with the
current client. If the script is run outside of tmux, it'll start a new
client on that terminal.

I remapped the `Prefix+f` keybind in tmux and the `CTRL+f` keybind in
zsh to that script.

This script and the ideas for the keybindings were copied from
ThePrimeagen's dotfile repository, plus a change from one of their
unmerged pull requests and some brief comments of my own.