vim: fix liquid front matter syntax in markdown
Markdown files with Liquid front matter (e.g. Jekyll pages) were parsed
erroneously by vim due to the `---` describing the end of the front
matter was parsed as a setext-style header.

This commit includes `.vim/after` scripts that include correct syntax
highlighting and a folding function.

As far as I know this is fixed in tpope's latest version of
vim-markdown, but NeoVim doesn't include it yet and I didn't want to add
another plugin for something so trivial. I found the code for these
changes on:

https://habamax.github.io/2019/03/07/vim-markdown-frontmatter.html

So thank you, habamax!