nvim: improved colorscheme
Certain highlights in the `tokyonight` colorscheme are illegible
(probably by design), esp. the line numbers on the gutter. As I use
relative numbers to move around a lot, this was impractical and so I
changed it.

But by doing so I made folded text illegible.

I should probably, in the future, change the highlights related to line
numbers in the gutter directly instead of modifying `fg_gutter`.